Bugfix for api.json not exists

+++ User.py
@@ -121,7 +121,8 @@ class Users:
         destination = Path(f'{self.api_backup}/api-keys_{date}.json')
         if not self.api_backup.exists():
             self.api_backup.mkdir(parents=True)
-        shutil.copy(PurePath(self.api_path), destination)
+        if Path(self.api_path).exists():
+            shutil.copy(PurePath(self.api_path), destination)
         with Path(f'{self.api_path}').open("w", encoding="UTF-8") as f:
             json.dump(save_users, f, indent=4)
